split nim_action oxo_action
